Style : CommentDao 코드리뷰 수정

+++ src/main/java/project/ffboard/service/CommentServiceImpl.java
@@ -15,7 +15,7 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
-    public Long addComment(Comment comment) {
+    public int addComment(Comment comment) {
         return commentDao.addComment(comment);
     }
